Extract main import.go logic into dbutil.go
